+++ R/expand.R
@@ -25,6 +25,9 @@
 #'
 #'   To fill in values that are missing altogether, use expressions like
 #'   \code{year = 2010:2020} or \code{year = \link{full_seq}(year)}.
+#'
+#'   Length-zero (empty) elements are automatically dropped.
+#' @param x For \code{nesting_} and \code{crossing_} a list of variables.
 #' @seealso \code{\link{complete}} for a common application of \code{expand}:
 #'   completing a data frame with missing combinations.
 #' @seealso \code{\link{expand_}} for a version that uses regular evaluation
@@ -120,20 +123,14 @@ expand_.grouped_df <- function(data, dots, ...) {
 #' @export
 #' @rdname expand
 crossing <- function(...) {
-  crossing_(list(...))
+  crossing_(tibble::lst(...))
 }
 
 #' @export
 #' @rdname expand
-#' @importFrom tibble data_frame
-nesting <- function(...) {
-  df <- data_frame(...)
-  df <- dplyr::distinct(df)
-  df[do.call(order, df), , drop = FALSE]
-}
-
 crossing_ <- function(x) {
   stopifnot(is.list(x))
+  x <- drop_empty(x)
 
   is_atomic <- vapply(x, is.atomic, logical(1))
   is_df <- vapply(x, is.data.frame, logical(1))
@@ -146,7 +143,7 @@ crossing_ <- function(x) {
     )
   }
 
-  # turn each data frame into single column data frame
+  # turn each atomic vector into single column data frame
   col_df <- lapply(x[is_atomic], function(x) data_frame(x = ulevels(x)))
   col_df <- Map(setNames, col_df, names(x)[is_atomic])
   x[is_atomic] <- col_df
@@ -160,3 +157,26 @@ cross_df <- function(x, y) {
   dplyr::bind_cols(x[x_idx, , drop = FALSE], y[y_idx, , drop = FALSE])
 }
 
+
+#' @export
+#' @rdname expand
+#' @importFrom tibble data_frame
+nesting <- function(...) {
+  nesting_(tibble::lst(...))
+}
+
+#' @export
+#' @rdname expand
+nesting_ <- function(x) {
+  stopifnot(is.list(x))
+  x <- drop_empty(x)
+
+  df <- as_data_frame(x)
+  df <- dplyr::distinct(df)
+  df[do.call(order, df), , drop = FALSE]
+}
+
+drop_empty <- function(x) {
+  empty <- vapply(x, function(x) length(x) == 0, logical(1))
+  x[!empty]
+}