+++ tests/testthat/test-expand.R
@@ -59,3 +59,12 @@ test_that("preserves ordered factors", {
 
   expect_equal(df$a, ordered("a"))
 })
+
+
+test_that("zero length inputs are automatically dropped", {
+  tb <- tibble::tibble(x = 1:5)
+
+  expect_equal(expand(tb, x, y = numeric()), tb)
+  expect_equal(nesting(x = tb$x, y = numeric()), tb)
+  expect_equal(crossing(x = tb$x, y = numeric()), tb)
+})