+++ R/utils.R
@@ -24,21 +24,12 @@ compact <- function(x) x[vapply(x, length, integer(1)) > 0]
 
 #' Extract numeric component of variable.
 #'
-#' This uses a regular expression to strip all non-numeric character from
-#' a string and then coerces the result to a number. This is useful for
-#' strings that are numbers with extra formatting (e.g. $1,200.34).
+#' DEPRECATED: please use \code{readr::parse_numeric()} instead.
 #'
 #' @param x A character vector (or a factor).
 #' @export
-#' @examples
-#' extract_numeric("$1,200.34")
-#' extract_numeric("-2%")
-#'
-#' # The heuristic is not perfect - it won't fail for things that
-#' # clearly aren't numbers
-#' extract_numeric("-2-2")
-#' extract_numeric("12abc34")
 extract_numeric <- function(x) {
+  message("extract_numeric() is deprecated: please use readr::parse_numeric() instead")
   as.numeric(gsub("[^0-9.-]+", "", as.character(x)))
 }