+++ NEWS.md
@@ -1,5 +1,10 @@
 # tidyr 0.5.1.9000
 
+## New functions
+
+* `drop_na()` removes observations which have `NA` in the given variables. If no
+  variables are given, all variables are considered (#194, @janschulz).
+
 # tidyr 0.5.1
 
 * Restored compatibility with R < 3.3.0 by avoiding `getS3method(envir = )` (#205, @krlmlr).