+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidyr 0.5.1.9000
 
+* `full_seq()` works correctly for dates and date/times.
+
 # tidyr 0.5.1
 
 * Restored compatibility with R < 3.3.0 by avoiding `getS3method(envir = )` (#205, @krlmlr).