+++ NEWS.md
@@ -7,7 +7,8 @@
   that take a list as input.
 
 * Renamed `table4` and `table5` to `table4a` and `table4b` to make their
-  connection more clear.
+  connection more clear. The `key` and `value` variables in `table2` have 
+  been renamed to `type` and `cases`.
 
 * `full_seq()` works correctly for dates and date/times.