+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidyr 0.6.0.9000
+
 # tidyr 0.6.0
 
 ## API changes